+++ doc/whats_new.rst
@@ -4,6 +4,16 @@
 NEW
 ---
 
+- :func:`nilearn.datasets.load_mni152_gm_template` takes the skullstripped
+  1mm-resolution version of the gray-matter MNI152 template and re-samples it
+  using a different resolution, if specified.
+- :func:`nilearn.datasets.load_mni152_wm_template` takes the skullstripped
+  1mm-resolution version of the white-matter MNI152 template and re-samples it
+  using a different resolution, if specified.
+- :func:`nilearn.datasets.load_mni152_gm_mask` loads mask from the gray-matter
+  MNI152 template.
+- :func:`nilearn.datasets.load_mni152_wm_mask` loads mask from the white-matter
+  MNI152 template.
 - :func:`nilearn.signal.clean` accepts new parameter `sample_mask`.
   shape: (number of scans - number of volumes removed, )
   Masks the niimgs along time/fourth dimension to perform scrubbing (remove
@@ -40,6 +50,24 @@ Fixes
 Enhancements
 ------------
 
+- :func:`nilearn.datasets.load_mni152_template` resamples now the template to
+  a preset resolution different from the resolution of the original template,
+  i.e. 1mm. The default resolution is 2mm, which means that the new template is
+  resampled to the resolution of the old template. Nevertheless, the shape of
+  the template changed from (91, 109, 91) to (99, 117, 95); the affine also
+  changed from array([[-2., 0., 0., 90.], [0., 2., 0., -126.],
+  [0., 0., 2., -72.], [0., 0., 0., 1.]]) to array([[1., 0., 0., -98.],
+  [0., 1., 0., -134.], [0., 0., 1., -72.], [0., 0., 0., 1.]]). Additionally,
+  the new template has also been rescaled; whereas the old one contained int16
+  floats varying between 0 and 8339, the new one contains floats varying between
+  0. and 1.
+- :func:`nilearn.datasets.load_mni152_brain_mask` accepts now the parameter
+  resolution, which will set the resolution of the template used for the
+  masking.
+- :func:`nilearn.masking.compute_brain_mask` accepts now as input the
+  whole-brain, 1mm-resolution, MNI152 T1 template instead of the averaged,
+  whole-brain, 2mm-resolution MNI152 T1 template; it also accepts as input the
+  gray-matter and white-matter ICBM152 1mm-resolution templates dated from 2009.
 - :func:`nilearn.plotting.view_markers` now accepts an optional argument `marker_labels` to provide labels to each marker.
 - :func:`nilearn.plotting.plot_surf` now accepts new values for `avg_method` argument, such as `min`, `max`, or even a custom python function to compute the value displayed for each face of the plotted mesh.
 - :func:`nilearn.plotting.view_img_on_surf` can now optionally pass through